pipe command output
