+++ src/main.rs
@@ -1,7 +1,6 @@
 use affected::logger::init_logger;
-use affected::{
-    get_project, list_affected_files, list_affected_projects, run_task_by_name, Config,
-};
+use affected::tasks;
+use affected::{get_project, list_affected_files, list_affected_projects, Config};
 use anyhow::{bail, Result};
 use clap::{Parser, Subcommand};
 use git2::Repository;
@@ -119,7 +118,7 @@ fn main() -> Result<()> {
             }
         },
         Commands::Run { task } => {
-            run_task_by_name(&workspace_root, &repo, &config, task)?;
+            tasks::run_task_by_name(&workspace_root, &repo, &config, task)?;
             println!("Done");
         }
     }
